use a videoRef instead of HTMLVideoElement
The HLSSource was struggling with the video element directly and was destroying and reloading itself. Shifting to a ref stabilized it. Probably better ways to approach it, but this works for now.

Also fixed an issue with seeking by sending and explicit END_SEEK event that sets the seekingTime to 0. Otherwise, the seekingTime value lingered around and was causing oddball behavior.

state made it much easier to sleuth this down.

also leaned into centralizing the state selectors after a nice affirmation from Matt Pocock that this approach is solid https://twitter.com/mpocock1/status/1434761800520814592?s=20