feat(testingaccessibility): bring player into TA application
wanted to see about getting mux integrated and ran into a few issues with the player. On was some typing disparities with useSelector that I had to ignore and don't fully understand. I was also required to implement "loadResource" in the provider, which I think shouldn't be the case. should let us just run a video with a essentially just the source URL imo.