feat(tt): updates exercise query for video resource
fairly sweeping change with some minor updates to sanity, but the structure of the exercise query has changed to normalize properties like `stackblitz` and `muxPlaybackId` as properties of the exercise versus a sub-object.

the issue was that solution video resources were a reference and harder to get at, so this flattens everything to make it simpler.