cleanup: ignore sitemap files from git
The sitemap files should only be needed as build assets in Vercel. We
shouldn't need to keep them under version control. They create a lot of
churn and noise in the git working directory, commits, and PRs. By
adding them to the `.gitignore` for the monorepo, we can get rid of that
clutter.