build(ts): allow build when missing algolia key
The build fails locally when the ALGOLIA_API_WRITE_KEY is missing. It
fails in an obscure way that makes it difficult to get the app up and
running.

This commit moves the key to a more appropriate file
(`env.local.template`) and then updates the Algolia client code to
fallback to an empty string if it is missing.