feat(ta): validate algolia key per-environment
Use Zod to parse and validate the ALGOLIA_API_WRITE_KEY env var which
smooths out the potentially undefined value and deals with type
inference. In the case that this value is undefined in the context of
the production environment, it raises a Zod validation error. This will
happen during the build if that value is missing which is definitely
when we want to catch it.