fix(ta): prevent default export of server-only code
The server-side-only algolia client configured with the secret API write
key should only be imported into API code. Because it was being `*`
exported from the `utils/algolia/index.ts` file, it was being bundled
into the frontend code.

This was originally causing silent errors because the algolia client was
being configured with a blank value (that write key is only visible
servers-side). The schema parsing -- meant only to validate the key
server-side -- that I recently added was being triggered on the frontend
which caused that aforementioned error to fully surface.

This commit removes the re-export of the server-side algolia client.
Instead, the API code can import it directly from
`utils/algolia/sanity-algolia-client.ts`. Having to explicitly import it
form this location should help avoid this code getting accidentally
included in the frontend bundle going forward.