build(ta): remove `build:dev` from TA pnpm scripts
According to a recent discussion, the `build:dev` command (introduced in
https://github.com/skillrecordings/products/commit/bf0fbe34244cd6414352b6f1308e05010313ecf1#diff-284649712bae897eb9f7e577e57534aa1ec714618f8b0b4641d70de63818c8ff)
is not used. Instead, if you want to build the app, you should run `pnpm
build`.

This commits removes `build:dev` from both `package.json` and the
`README.md` to avoid confusion about how to set up and build this app.