feat(db): spike out data migration script setup
This is a WIP/spike on how to write a data migration script for a
scenario where the schema has been changed and now data needs to be
migrated to that updated schema.

It should lean on Prisma types and should be runnable across the various
apps using the schema for their databases.

It uses `ts-node` to execute the typescript data migration file. Open to
suggestions if there is something else I should be using for this
particular piece of this. It's what I found.

Possibly my brain is fried from a long day, but I'm struggling to
understand where the node environment is being derived from when I
execute this script. No matter what directory I execute this from,
`ts-node` has the impression that `DATABASE_URL` is set to the URL
defined in TA's `.env` file. What am I missing? How do I make the
environment variables specific to the project this script is being run
in?

I'm executing it with `npx ts-node
../../packages/database/migrate-data/<name-of-file>.ts`.