feat(ta): add one-to-many relation for bulk coupons
This PR represents the second and third step of the Expand and Contract
pattern for replacing the existing one-to-one relation between Bulk
Coupon and Bulk Purchase with a one-to-many relation of Bulk Coupon to
Purchases.

Before this can be merged, the schema needs to be fully deployed to the
existing production database environments. Those Deploy Requests have
already been opened at:
1. https://app.planetscale.com/skill-recordings/testing-accessibility/deploy-requests/14
2. https://app.planetscale.com/skill-recordings/total-typescript/deploy-requests/3