feat(ta): allow Individual to upgrade to Team
This adds a page that a logged in user with a purchase can visit. It has
a 'Buy More Seats' component that takes the user to a Stripe Checkout
Session. Once they complete the purchase, a new Bulk Coupon will be
created for the user with that many seats. It doesn't impact their
existing individual purchase, which we maintain for record keeping and
to determine this user's access.