feat(commerce-server): add fn for purchase type
This adds a dedicated function for computing the purchase type for a
checkout session. This can be used by apps to determine what type of
purchase has just been made (e.g. a new bulk coupon purchase) so that we
can trigger the appropriate post-purchase actions and display copy
specific to each purchase scenario.