feat(tt): add quick-and-dirty logout page
I needed a way to log out of an account while building and smoke testing
features. So, I added a `/logout` page with a button that uses
NextAuth's `signOut()` function.

We'll eventually need a more robust logout mechanism. But this will work
for now.

A noteable issue I ran into while implementing this (which I was unable
to debug) is that the button redirects to `localhost:3000` instead of
`localhost:3016`. I've explicitly passed in a full redirect URL, I've
ensure there aren't any env vars set that could be interfering. Beyond
that, I don't have enough expose to NextAuth to know why it would be
redirect to a completely different port.