fix(db): use same ID for bulk coupon and purchase
This was an oversight during a refactor. Because the `id` on a `Coupon`
is auto-generated if not given, this was silenting failing by creating a
Coupon with one `id` and a Purchase with a different `bulkCouponId`.
This is a situation where it would be really nice to have foreign key
constraints.