refactor(tt): experiment with better page types
Use Next.js type helpers to get better typing across the
`getServerSideProps` and rendered component boundaries. It is not
perfect and will get more out of `tRPC` as we continue to expand how we
are using that. Nevertheless, we have a lot of pages that us
`getServerSideProps` and this can help improve our typings for those
situations.