feat: improve 'out of seats' state of Team page
This gets rid of the old 'buy more seats' button that conditionally
displays when there are no more seats. We can instead rely on the 'Buy
More Seats' widget below this that is always visible.

It also adds a disabled attribute and disabled styles to the various
team buttons, instead of hiding them, to make it clearer what is
available if you get more seats.