use skill-lesson in pro tailwind
this is using the `skill-lesson` package for metadata and media loading that makes use of `LessonResource` base type instead of local `Exercise` where the media is embedded in the type.

The embedded media caused an issue because it varies from product to product so instead we separate them for runtime loading.

`sandpack` and `stackblitz` are an example of clear difference.

![different](https://media.giphy.com/media/bJQTg7z8WrQRzVSavV/giphy-downsized.gif)