feat(sks): avoid confusing PPP logic
The logic to not create a promo code when there is a PPP coupon and the
quantity is greater than 1 is hard to reason about. Instead, we can
null-out the merchant coupon earlier up in the function if it is PPP and
the quantity is greater than 1. That clears things up down the rest of
the function.