feat(kcd): add mux and sanity import scripts
Once we have a Mux environment for Epic React, the Mux asset upload
script can be run. This will produce a JSON output with all the Mux
video identifiers.

The Sanity import script is still somewhat WIP in that it needs to be
run against a non-prod Sanity environment and evaluated for correctness,
but it is probably 95% of the way there. The Mux JSON output is needed
to do that.