feat(kcd): add script with re-runnability built-in
This script synchronously transfers records from a local LessonProgress
table to a remote Planetscale LessonProgress table. Because of
consistent connection failures during the multi-threaded streaming
approach, I decided to go another direction with this. With this script,
we always know how far we've processed through the table of records, so
if the connection gets closed, we can record where we are, and pick up
where we left off. This makes the script idempotent which means that it
can be run outside of the cutover processes after the site has been
brought back up. The original import script will still need to be run
first in order to load those records up into a local Docker database.