feat(kcd): script attempts to re-run itself
If a connection gets closed, which happens quite often with planetscale,
then this script will note where it left off and re-run itself from that
point. In this way, the script doesn't need constant babysitting to
constantly start the script when it exits early. However, if it is
failing over and over without making any progress, then the script will
notice that and exit with a message.