jb/refactor coupon application in pricing logic (#1093)
* feat(commerce): move pppApplied into helper

* feat(commerce): move hasPPP check to helper

* feat(commerce): return result instead of throwing

* feat(commerce): move more PPP logic into helper

* feat(commerce): consolidate PPP logic

* cleanup(commerce): remove user purchases lookup

* feat(commerce): combine PPP formatting scenarios

* feat(commerce): move bulk coupon logic into helper

* feat(commerce): try renaming incoming coupon

The incoming `merchantCouponId` is a reference to a candidate coupon,
e.g. maybe a site-wide coupon. We may override this with a better
coupon, so we should treat it is as a candidate while we figure out what
coupon will actually be applied.

* feat(commerce): make coupon comparison more explicit

* feat(commerce): find available coupons in helper

* feat(commerce): lookup bulk coupon in helper

* cleanup(commerce): remove unused PPP percent const

* feat(commerce): apply bulk, combine conditions

* feat(commerce): tighten fixed discount logic

* feat(commerce): pull upgradeDetails logic up

* feat(commerce): collapse bulk logic into rest

* feat(commerce): collapse available PPP logic

* feat(commerce): collapse extraneous 'special' logic

The block toward the top of the price formatting that was catching a
first pass of 'special' coupon logic is now redundant with the following
logic that applies any coupons, including 'special', as needed.

* feat(commerce): remove commented error-throwing

* cleanup(commerce): clean up some comments

* cleanup(commerce): more coupon comment cleanup

* feat(commerce): reduce PPP details logic

* feat(commerce): remove unused hasPurchaseWithPPP

* feat(commerce): note future refactoring

* feat(commerce): get bulk coupon, instead of details

* feat(commerce): no need to return pppDetails

The `pppDetails` served their purpose for the refactoring, but all that
info is self-contained to the helper module and doesn't need to be
passed to the price formatter anymore.

* feat(commerce): remove coupon `code` logic branch

The `code` to `Coupon.id` look up and translation happens in the
`propsForCommerce` endpoint which then only passes along a coupon ID
after that point. There is no consumer or app of the
`formatPricesForProduct` code path that passes in a `code`.

* feat(commerce): remove last redundant price block

* fix(commerce): don't lookup coupon w/ undefined

This is sorta weird Prisma behavior, if the main attribute of the where
clause is `undefined` in a `findFirst` query, it will exclude that
condition and then it is just going to grab the first row from the table
and return that. https://github.com/prisma/prisma/issues/5149

* fix(commerce): ensure PPP does not leak through

In the scenario where you had 'checked' the PPP discount box and then
changed to Team/Bulk purchase (via toggle), the applied PPP merchant
coupon would be carried over and applied as a fall-through value.

This updates the logic to ensure PPP only gets assigned via the PPP code
path and that special coupons are the ones that fall through.

* fix(commerce): remove unused var, function call

* feat(commerce): handle PPP + Bundle Upgrade

This deals with linear issue TT-331