feat(commerce): indicate bulk purchase in price
The FormattedPrice should indicate to consumers whether the purchase is
going to be a bulk purchase or not. This is because the price formatting
logic does a qualified seat count lookup and knows if the purchase is
individual or adding to an existing bulk purchase.

Existing logic might think a purchase of `quanity = 1` is an individual
purchase when in fact a single seat is being added to an existing bulk
purchase.