feat(commerce): get pricing w/o PPP auto-applied
This adds an option to the Pricing Formatter to request the pricing with
or without PPP being auto-applied. For TT, this means that we can
initially auto-apply PPP if we know a user's previous purchase was PPP
and then if they interact with the toggle, then we can switch off the
auto-apply.