feat(commerce): record upgradedFromId on purchase
We have the `upgradedFromId` on the `Purchase` table which exists to
record when a Purchase record is one that was upgraded from an existing
Purchase record. We haven't been using it tho.

This includes that info in the Stripe metadata during the checkout phase
when we have easy access to the purchase being upgraded from and then in
the post-purchase handler, we record that on the new purchase record.