feat(commerce): multi-purchase fixed discount
When we compute the fixed discount for a PPP to Unrestricted upgrade, we
have been assuming that the _Purchase to be Upgraded_ is the only
purchase. However, it is possible that someone purchase Core(PPP), then
purchased an upgrade to Bundle(PPP), and now wants to upgrade to
Unrestricted Bundle. In that case, we need to get the _amount paid_ for
each of the purchases in the chain of purchases when computing the fixed
discount, otherwise we are overcharging the user.