feat(tjs): remove /playlists index page + links
The `/playlists` index page was missing some logic to correctly display
the modules based on the user's purchase. I noticed that the existing
TJS site doesn't have/support a playlists index page. It instead only
shows individual modules. So, intead of updating the logic for this
extraneous index page, I'm going to remove it and the nav links to it.