feat(ew): handle unfound exercise scenario (#1305)
Because we look up the exercise based on a slug that comes from a URL
(user input), it is possible that the slug is bad and we don't find an
exercise. At the moment, the Zod parse fails the validation and the
server responds with a 500. Instead, we can handle the Record Not Found
scenario by returning `null` and having the various call sites react
appropriately. Often that means rendering a 404 response.

Co-authored-by: Vojta Holik <vojta@egghead.io>