fix: default Merchant entities to active status
When we are first setting up a Product and its pricing in Sanity, a
sanity callback fires which is handled by an inngest function. That
inngest function sets up records in Stripe and Planetscale. The
MerchantProduct and MerchantPrice records that it sets up in Planetscale
were not being given `status` values, so it was using the database
defaults of `0`. Since we are doing the initial setup of the product,
even if the product itself isn't active, we still want the Merchant
entities to be given a status of `1`. That indicates they are the active
values for this product once it is ready to go live.