feat: exclude coupon if not for productId
there are many pages where we show pricing for multiple products (e.g.
when there are different tiers) and the coupon coming from a coupon code
is only relevant to one of those products. Ensure that the price gets
calculated correctly for each by restricting the coupon to its product.