feat(commerce): allow PPP Upgrade in Pricing
The commerce logic itself was sending everything we needed for offering
a PPP Upgrade, but the display logic was not accounting for it (except
for the Blocked Overlay which didn't pass in a `purchased` prop). In the
event that the current product is available for a PPP Upgrade, the
Pricing component now overrides `purchased` to be `false` allowing the
UI logic to display the upgrade price and fixed discount amount.