fix(commerce): ensure discount amount is integer
Stripe requires the `amount_off` of a coupon to be an integer expressed
in cents. We were attempting to do that with `* 100`, but in this case
we had a float value that with a trailing series of decimals. This
now rounds it so that we are sure to have an integer.